Remove redudant casting

+++ app/src/org/commcare/android/view/DotsDetailView.java
@@ -190,7 +190,7 @@ public class DotsDetailView {
                 LinearLayout.LayoutParams params = new LinearLayout.LayoutParams(LinearLayout.LayoutParams.WRAP_CONTENT,LinearLayout.LayoutParams.FILL_PARENT);
                 params.leftMargin = 10;
                 params.rightMargin = 10;
-                params.weight = (float)(1f / day.boxes().length);
+                params.weight = 1f / day.boxes().length;
                 container.addView(details, params);
                 container.setOrientation(LinearLayout.HORIZONTAL);
             } else {
